# cedar-resource-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Better Metadata - Better Science",
    "version" : "2.8.3",
    "title" : "CEDAR Resource Server API"
  },
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/template-instances" : {
      "post" : {
        "tags" : [ "Template Instances" ],
        "summary" : "Create a template instance",
        "description" : "Create a template instance.",
        "operationId" : "createTemplateInstance",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier. The artifact will be created in this folder. The user must have write access to the folder. If not provided, the artifact will be created in the user's home folder.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template_instance",
          "description" : "The template instance to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TemplateInstance"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TemplateInstance"
            }
          },
          "201" : {
            "description" : "A template instance",
            "schema" : {
              "$ref" : "#/definitions/TemplateInstance"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-instances/{template_instance_id}" : {
      "get" : {
        "tags" : [ "Template Instances" ],
        "summary" : "Get a template instance",
        "description" : "Get a template instance.",
        "operationId" : "findTemplateInstance",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier. Example: https://repo.metadatacenter.org/template-instances/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "format",
          "in" : "query",
          "description" : "Output format type to display the content of the template instance. The allowed values are: 'jsonld', 'json', 'rdf-nquad'.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A template instance",
            "schema" : {
              "$ref" : "#/definitions/TemplateInstance"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Template Instances" ],
        "summary" : "Update a template instance",
        "description" : "Update a template instance.",
        "operationId" : "updateTemplateInstance",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template_instance",
          "description" : "The template instance to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TemplateInstance"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A template instance",
            "schema" : {
              "$ref" : "#/definitions/TemplateInstance"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Template Instances" ],
        "summary" : "Delete a template instance",
        "description" : "Delete a template instance.",
        "operationId" : "deleteTemplateInstance",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation (no content)"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-instances/{template_instance_id}/details" : {
      "get" : {
        "tags" : [ "Template Instances", "Resource Details" ],
        "summary" : "Get details of a template instance",
        "description" : "Get details of a template instance.",
        "operationId" : "findTemplateInstanceDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-instances/{template_instance_id}/download" : {
      "post" : {
        "tags" : [ "Template Instances" ],
        "summary" : "Download a template instance",
        "description" : "Download a template instance as JSON or YAML, selected via the Accept header.",
        "operationId" : "downloadTemplateInstance",
        "produces" : [ "application/json", "application/x-yaml" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Accept",
          "in" : "header",
          "description" : "Desired output format: 'application/json' or 'application/yaml'.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "compact",
          "in" : "query",
          "description" : "When downloading YAML, produce a compact representation.",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "The template instance content as an attachment"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-instances/{template_instance_id}/permissions" : {
      "get" : {
        "tags" : [ "Template Instances", "Permissions" ],
        "summary" : "Get permissions of a template instance",
        "description" : "Get permissions of a template instance.",
        "operationId" : "getTemplateInstancePermissions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Template Instances", "Permissions" ],
        "summary" : "Update permissions of a template instance",
        "description" : "Update permissions of a template instance.",
        "operationId" : "updateTemplateInstancePermissions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-instances/{template_instance_id}/report" : {
      "get" : {
        "tags" : [ "Template Instances", "Resource Report" ],
        "summary" : "Get report of a template instance",
        "description" : "Get report of a template instance.",
        "operationId" : "getTemplateInstanceReport",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_instance_id",
          "in" : "path",
          "description" : "Template Instance identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
    "/templates/{template_id}" : {
      "get" : {
        "tags" : [ "Templates" ],
        "summary" : "Get a template",
        "description" : "Get a template.",
        "operationId" : "findTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier. Example: https://repo.metadatacenter.org/templates/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A template",
            "schema" : {
              "$ref" : "#/definitions/Template"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Templates" ],
        "summary" : "Update a template",
        "description" : "Update a template.",
        "operationId" : "updateTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template",
          "description" : "The template to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Template"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A template",
            "schema" : {
              "$ref" : "#/definitions/Template"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Templates" ],
        "summary" : "Delete a template",
        "description" : "Delete a template.",
        "operationId" : "deleteTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation (no content)"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/templates/{template_id}/details" : {
      "get" : {
        "tags" : [ "Templates", "Resource Details" ],
        "summary" : "Get details of a template",
        "description" : "Get details of a template.",
        "operationId" : "findTemplateDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "api_key" : {
      "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Template" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific template."
        }
      },
      "description" : "A CEDAR template (open JSON-LD document)."
    },
...
    "TemplateInstance" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific template instance."
        }
      },
      "description" : "A CEDAR template instance (open JSON-LD document)."
    },
...
  }
}